Added examples folder && update README.md
